+++ integrations/sentry/CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- towncrier release notes start -->
 
+
+## 0.1.98 (2024-12-12)
+
+
+### Improvements
+
+- Added support for Sentry users and teams
+- Updated issues to add an assignedTo relation
+
+
 ## 0.1.97 (2024-12-10)